// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/features/dpp-tapp-list/project-details/models/cal-financial-analysis.model.ts
export class CalFinancialAnalysis {

  fiscalYear: any;
  noOfFiscalYear: number = 0;
  capitalCost: number = 0;
  operatingCost: number = 0;
  totalCost: number = 0;

  discValueOfTotalCost1: number = 0;
  benifit1: number = 0;
  discValueOfBenifit1: number = 0;

  discValueOfTotalCost2: number = 0;
  benifit2: number = 0;
  discValueOfBenifit2: number = 0;

}